Update delete address phone page
