+++ server/routes/contacts/manage/addresses/delete-address-phone/manageContactDeleteAddressPhoneController.ts
@@ -1,25 +1,20 @@
 import { Request, Response } from 'express'
 import { Page } from '../../../../../services/auditService'
 import { PageHandler } from '../../../../../interfaces/pageHandler'
-import ReferenceCodeType from '../../../../../enumeration/referenceCodeType'
-import ReferenceDataService from '../../../../../services/referenceDataService'
-import { PhoneNumberSchemaType } from '../phoneSchemas'
 import { ContactsService } from '../../../../../services'
 import { Navigation } from '../../../common/navigation'
 import ContactDetails = contactsApiClientTypes.ContactDetails
 import ContactAddressDetails = contactsApiClientTypes.ContactAddressDetails
 import ContactAddressPhoneDetails = contactsApiClientTypes.ContactAddressPhoneDetails
+import Urls from '../../../../urls'
 import { FLASH_KEY__SUCCESS_BANNER } from '../../../../../middleware/setUpSuccessNotificationBanner'
 import { formatNameFirstNameFirst } from '../../../../../utils/formatName'
-import Urls from '../../../../urls'
+import { getUpdateAddressDetails } from '../common/utils'
 
-export default class ManageContactEditAddressPhoneController implements PageHandler {
-  constructor(
-    private readonly contactsService: ContactsService,
-    private readonly referenceDataService: ReferenceDataService,
-  ) {}
+export default class ManageContactDeleteAddressPhoneController implements PageHandler {
+  constructor(private readonly contactsService: ContactsService) {}
 
-  public PAGE_NAME = Page.EDIT_ADDRESS_PHONE_PAGE
+  public PAGE_NAME = Page.DELETE_ADDRESS_PHONE_PAGE
 
   GET = async (
     req: Request<{
@@ -31,12 +26,8 @@ export default class ManageContactEditAddressPhoneController implements PageHand
     }>,
     res: Response,
   ): Promise<void> => {
-    const { user } = res.locals
     const { prisonerNumber, contactId, prisonerContactId, contactAddressId, contactAddressPhoneId } = req.params
-    const contact: ContactDetails = await this.contactsService.getContact(Number(contactId), user)
-    const address = contact.addresses.find(
-      (item: ContactAddressDetails) => item.contactAddressId === Number(contactAddressId),
-    )
+    const { address, formattedAddress } = await getUpdateAddressDetails(this.contactsService, req, res)
     const phone: ContactAddressPhoneDetails = address.phoneNumbers.find(
       (aPhone: ContactAddressPhoneDetails) => aPhone.contactAddressPhoneId === Number(contactAddressPhoneId),
     )
@@ -45,46 +36,38 @@ export default class ManageContactEditAddressPhoneController implements PageHand
         `Couldn't find phone with id ${contactAddressPhoneId} for contact ${contactId} and address ${contactAddressId}. URL probably entered manually.`,
       )
     }
-    const currentType = res.locals?.formResponses?.['type'] ?? phone.phoneType
-    const typeOptions = await this.referenceDataService.getReferenceData(ReferenceCodeType.PHONE_TYPE, user)
-    const navigation: Navigation = { backLink: Urls.editContactMethods(prisonerNumber, contactId, prisonerContactId) }
-    const viewModel = {
-      typeOptions,
-      phoneNumber: res.locals?.formResponses?.['phoneNumber'] ?? phone.phoneNumber,
-      type: currentType,
-      extension: res.locals?.formResponses?.['extension'] ?? phone.extNumber,
-      contact,
-      navigation,
-      address,
+    const navigation: Navigation = {
+      backLink: Urls.editContactMethods(prisonerNumber, contactId, prisonerContactId),
+      cancelButton: Urls.contactDetails(prisonerNumber, contactId, prisonerContactId, 'contact-methods'),
     }
-    res.render('pages/contacts/manage/contactMethods/addEditAddressPhone', viewModel)
+    res.render('pages/contacts/manage/contactMethods/address/phone/confirmDeleteAddressPhone', {
+      phone,
+      navigation,
+      formattedAddress,
+    })
   }
 
   POST = async (
-    req: Request<
-      {
-        prisonerNumber: string
-        contactId: string
-        prisonerContactId: string
-        contactAddressId: string
-        contactAddressPhoneId: string
-      },
-      unknown,
-      PhoneNumberSchemaType
-    >,
+    req: Request<{
+      prisonerNumber: string
+      contactId: string
+      prisonerContactId: string
+      contactAddressId: string
+      contactAddressPhoneId: string
+    }>,
     res: Response,
   ): Promise<void> => {
     const { user } = res.locals
     const { prisonerNumber, contactId, prisonerContactId, contactAddressId, contactAddressPhoneId } = req.params
-    const { phoneNumber, type, extension } = req.body
-    await this.contactsService.updateContactAddressPhone(
-      Number(contactId),
-      Number(contactAddressId),
-      Number(contactAddressPhoneId),
+    const contactIdNumber = Number(contactId)
+    const contactAddressIdNumber = Number(contactAddressId)
+    const contactPhoneIdNumber = Number(contactAddressPhoneId)
+
+    await this.contactsService.deleteContactAddressPhone(
+      contactIdNumber,
+      contactAddressIdNumber,
+      contactPhoneIdNumber,
       user,
-      type,
-      phoneNumber,
-      extension,
     )
     await this.contactsService
       .getContactName(Number(contactId), user)
