+++ server/routes/contacts/manage/manageContactsRoutes.ts
@@ -47,9 +47,9 @@ import AddressCheckAnswersController from './addresses/address-check-answers/add
 import RestrictionsService from '../../../services/restrictionsService'
 import ManageApprovedToVisitController from './relationship/approved-to-visit/manageApprovedToVisitController'
 import ManageRelationshipStatusController from './relationship/status/manageRelationshipStatusController'
-import ManageContactAddAddressPhoneController from './phone/add-address-phone/manageContactAddAddressPhoneController'
-import ManageContactEditAddressPhoneController from './phone/edit-address-phone/manageContactEditAddressPhoneController'
-import ManageContactDeleteAddressPhoneController from './phone/delete-address-phone/manageContactDeleteAddressPhoneController'
+import ManageContactAddAddressPhoneController from './addresses/add-address-phone/manageContactAddAddressPhoneController'
+import ManageContactEditAddressPhoneController from './addresses/edit-address-phone/manageContactEditAddressPhoneController'
+import ManageContactDeleteAddressPhoneController from './addresses/delete-address-phone/manageContactDeleteAddressPhoneController'
 import UsePrisonerAddressController from './addresses/use-prisoner-address/usePrisonerAddressController'
 import PrisonerAddressService from '../../../services/prisonerAddressService'
 import { PageHandler } from '../../../interfaces/pageHandler'
