--- server/views/pages/contacts/manage/contactMethods/confirmDeleteAddressPhone.njk
@@ -1,87 +1,0 @@
-{% extends "partials/layout.njk" %}
-{% from "partials/miniProfile/macro.njk" import miniProfile %}
-{% from "govuk/components/button/macro.njk" import govukButton %}
-{% from "govuk/components/summary-list/macro.njk" import govukSummaryList %}
-
-{% set pageTitle = applicationName + " - Delete address phone number" %}
-{% set title = "Are you sure you want to delete this phone number?" %}
-{% set mainClasses = "app-container govuk-body" %}
-
-{% block content %}
-
-    {% include 'partials/navigation.njk' %}
-    {{ miniProfile(prisonerDetails) }}
-    <span class="govuk-caption-l">Contacts</span>
-    <div class="govuk-grid-row">
-        <div class="govuk-grid-column-three-quarters">
-            <h1 class="govuk-heading-l" data-qa="main-heading">{{ title }}</h1>
-            {% set formattedAddress= {
-                flat: address.flat,
-                premise: address.property,
-                street: address.street,
-                area: address.area,
-                city: address.cityDescription,
-                county: address.countyDescription,
-                postalCode: address.postcode,
-                country: address.countryDescription
-            } %}
-            <p data-qa="address-reference">
-                {{ formattedAddress | addressToLines | escape | nl2br | safe }}
-            </p>
-            <form method='POST'>
-                <input type="hidden" name="_csrf" value="{{ csrfToken }}" />
-                {{ govukSummaryList({
-                    rows: [
-                        {
-                            key: {
-                                text: "Phone number"
-                            },
-                            value: {
-                                html: phone.phoneNumber,
-                                classes: 'phone-number-value'
-                            }
-                        },
-                        {
-                            key: {
-                                text: "Type"
-                            },
-                            value: {
-                                html: phone.phoneTypeDescription,
-                                classes: 'type-value'
-                            }
-                        },
-                        {
-                            key: {
-                                text: "Extension"
-                            },
-                            value: {
-                                html: phone.extNumber if phone.extNumber else 'Not provided',
-                                classes: 'extension-value'
-                            }
-                        }
-                    ]
-                }) }}
-
-                <div class="govuk-button-group">
-                    {{ govukButton({
-                        text: "Yes, delete",
-                        type: "submit",
-                        classes: 'govuk-!-margin-top-6 govuk-button--warning',
-                        attributes: {"data-qa": "continue-button"},
-                        preventDoubleClick: true
-                    }) }}
-                    {{ govukButton({
-                        text: "No, do not delete",
-                        type: "button",
-                        classes: 'govuk-!-margin-top-6 govuk-button--secondary',
-                        attributes: {"data-qa": "cancel-button"},
-                        href: navigation.cancelButton,
-                        preventDoubleClick: true
-                    }) }}
-                </div>
-            </form>
-
-        </div>
-    </div>
-
-{% endblock %}
